Re #274: Swap print for debug to honour log level
Also combined logging to a single line now we don't pretty-print
and removed double-logging for DMs (once in the block of all retrieved
DMs and once when they get faux-streamed).

But since that may hide non-"message_create" messages, we now log
the JSON earlier, before we drop unsupported messages.